Setup RabbitMQ for communicating with the executor (#22)
* Setup RabbitMQ for communicating with the executor

* Define the message contracts

* Call the SubmissionPublisher when the code is submitted

* Convert the message DTO to string when sending to the queue

* Use object mapper for listeners

* Save the results of the test cases returned by the executor

* Add test for updateStatus and updateTimeTakenAndMemoryUsed methods

+++ backend/src/main/java/com/online/judge/backend/model/TestCaseResult.java
@@ -55,8 +55,6 @@ public class TestCaseResult {
 	@Column(name = "execution_time_seconds", precision = 2)
 	private Double executionTimeSeconds;
 
-	@Column(name = "memory_used_mb", precision = 2)
-	private Double memoryUsedMb;
+	@Column(name = "memory_used_mb")
+	private Integer memoryUsedMb;
 }
-
-// TestCaseResultId moved to its own file for clarity and maintainability.
